MGDCTRS-1477 chore: rename components
This change renames the connector gallery to a selection list.  It
implements several recent design changes.  There's also a fix to an
issue with the cache and a bunch of improvements around how the selector
lays out on the page and handles page resizing.
