+++ locales/en/cos-ui.json
@@ -19,6 +19,7 @@
   "category-crm": "CRM",
   "category-database": "Database",
   "category-developer-tools": "Developer tools",
+  "category-featured": "Featured",
   "category-google": "Google",
   "category-integration-and-delivery": "Integration & delivery",
   "category-logging": "Logging",
