+++ src/ProofOfConcepts/ConnectorSelectionList/ConnectorSelectionListLayout.css
@@ -1,0 +1,3 @@
+.connector-selection-list__sidebar-content.pf-c-sidebar__content {
+  background-color: inherit;
+}
