+++ src/ProofOfConcepts/ConnectorSelectionList/ConnectorSelectionListToolbar.css
@@ -1,0 +1,3 @@
+h3.connector-selection-list__toolbar-category-title {
+  margin-right: 2em !important;
+}
