+++ src/ProofOfConcepts/ConnectorSelectionList/ConnectorSelectionListViewport.css
@@ -1,16 +1,10 @@
-.connector-types-gallery__viewport-container {
-  overflow: auto scroll;
+.connector-selection-list__viewport-container {
+  overflow: auto auto;
   scroll-behavior: smooth;
   -webkit-overflow-scrolling: touch;
   position: relative;
   padding: 15;
 }
-
-/*
-.connector-types-gallery__viewport-container-inner {
-    margin: -5px;
-}
-*/
 
 .pf-c-window-scroller .ReactVirtualized__VirtualGrid {
   overflow: visible !important;
