--- src/ProofOfConcepts/ConnectorTypesGallery/ConnectorTypesGallery.css
@@ -1,11 +1,0 @@
-.connector-types-gallery__sidebar-content.pf-c-sidebar__content {
-  background-color: inherit;
-}
-
-.connector-types-gallery__vertical-tabs {
-  margin-bottom: 0px;
-}
-
-.connector-types-gallery__search-input {
-  margin-top: 1.25em;
-}
