--- src/ProofOfConcepts/ConnectorTypesGallery/ConnectorTypesGallerySidePanel.css
@@ -1,7 +1,0 @@
-.connector-types-gallery__side-panel-search-container {
-  padding-bottom: 1.2em;
-}
-
-.connector-types-gallery__side-panel-viewport {
-  overflow: auto auto;
-}
